Move ExistingDrugMedicationRequiredException
